requirements.txt
torch>=2.0.0
torchvision>=0.15.0
ftfy
regex
tqdm
numpy
pandas
matplotlib
seaborn
scipy
PyYAML
git+https://github.com/openai/CLIP.git



